reorg sqlite example
