+++ examples/sqlite/src/web.rs
@@ -1,0 +1,6 @@
+pub use app::App;
+pub use auth::Credentials;
+
+mod app;
+mod auth;
+mod protected;
