+++ examples/sqlite/src/web/auth.rs
@@ -1,0 +1,90 @@
+use askama::Template;
+use axum::{
+    extract::Query,
+    response::{IntoResponse, Redirect},
+    routing::{get, post},
+    Form, Router,
+};
+use http::StatusCode;
+use serde::Deserialize;
+
+use crate::users::AuthSession;
+
+#[derive(Template)]
+#[template(path = "login.html")]
+pub struct LoginTemplate {
+    message: Option<String>,
+    next: Option<String>,
+}
+
+// This allows us to extract the "next" field from the query string. We use this
+// to redirect after log in.
+#[derive(Debug, Deserialize)]
+pub struct NextUrl {
+    next: Option<String>,
+}
+
+// This allows us to extract the authentication fields from forms. We use this
+// to authenticate requests with the backend.
+#[derive(Debug, Clone, Deserialize)]
+pub struct Credentials {
+    pub username: String,
+    pub password: String,
+    next: Option<String>,
+}
+
+pub fn router() -> Router<()> {
+    Router::new()
+        .route("/login", post(self::post::login))
+        .route("/login", get(self::get::login))
+        .route("/logout", get(self::get::logout))
+}
+
+mod post {
+    use super::*;
+
+    pub async fn login(
+        mut auth_session: AuthSession,
+        Form(creds): Form<Credentials>,
+    ) -> impl IntoResponse {
+        let user = match auth_session.authenticate(creds.clone()).await {
+            Ok(Some(user)) => user,
+            Ok(None) => {
+                return LoginTemplate {
+                    message: Some("Invalid credentials.".to_string()),
+                    next: creds.next,
+                }
+                .into_response()
+            }
+            Err(_) => return StatusCode::INTERNAL_SERVER_ERROR.into_response(),
+        };
+
+        if auth_session.login(&user).await.is_err() {
+            return StatusCode::INTERNAL_SERVER_ERROR.into_response();
+        }
+
+        if let Some(ref next) = creds.next {
+            Redirect::to(next).into_response()
+        } else {
+            Redirect::to("/").into_response()
+        }
+    }
+}
+
+mod get {
+    use super::*;
+
+    pub async fn login(Query(NextUrl { next }): Query<NextUrl>) -> LoginTemplate {
+        LoginTemplate {
+            message: None,
+            next,
+        }
+    }
+
+    pub async fn logout(mut auth_session: AuthSession) -> impl IntoResponse {
+        match auth_session.logout() {
+            Ok(_) => Redirect::to("/login").into_response(),
+            Err(_) => StatusCode::INTERNAL_SERVER_ERROR.into_response(),
+        }
+    }
+}
