+++ examples/sqlite/src/web/protected.rs
@@ -1,0 +1,30 @@
+use askama::Template;
+use axum::{response::IntoResponse, routing::get, Router};
+use http::StatusCode;
+
+use crate::users::AuthSession;
+
+#[derive(Template)]
+#[template(path = "protected.html")]
+struct ProtectedTemplate<'a> {
+    username: &'a str,
+}
+
+pub fn router() -> Router<()> {
+    Router::new().route("/", get(self::get::protected))
+}
+
+mod get {
+    use super::*;
+
+    pub async fn protected(auth_session: AuthSession) -> impl IntoResponse {
+        match auth_session.user {
+            Some(user) => ProtectedTemplate {
+                username: &user.username,
+            }
+            .into_response(),
+
+            None => StatusCode::INTERNAL_SERVER_ERROR.into_response(),
+        }
+    }
+}
